requirements.txt
# Core dependencies
flask>=2.0.1
requests>=2.26.0
PyYAML>=5.4.1
python-dateutil>=2.8.2
gunicorn>=20.1.0
werkzeug>=2.0.0
markupsafe>=2.0.0
flask-login>=0.6.2
python-ldap>=3.4.3

# Development dependencies (not needed in production)
python-dotenv>=0.19.0
